Permit expressions which "obviously" resolve to Builtin.Int1 to be used directly as conditions.
Currently this is basically just references to variables;
calls to builtin functions returning Builtin.Int1 aren't obvious
enough.

I claim that it's okay for this to be a little hacky because it's
dealing with builtin types.  If these were user-facing, this would
need to be integrated into the constraint somehow.

Swift SVN r7737