IRGen: Add a 'loadForSwitch' method to UnionImplStrategy.
It's always safe to load the binary pattern of a union for switching, even if the union is address-only, so rather than have another code path for address-only dispatch, we can just break it down to a load + switch. loadForSwitch is usually equivalent loadAsTake, but doesn't need an is-loadable sanity check or the side-table magic for weak references.

Swift SVN r7794