[stdlib] StringDesign.rst: More Work In Progress (WIP)
Beginning to move all ObjC/Cocoa references out of the first part of
the document, so the high-level string design stands on its own.  It's
becoming clear that rationales need to be moved out of the sidebar and
back into the main body of the text, so checkpointing before
restructuring.

Swift SVN r7899