IRGen: Add TypeInfo methods for dynamic extra inhabitant witnesses.
For non-fixed-layout union types, we want value witness functions for the union members that map extra inhabitants to integer indices, so that they can be used by generic union dispatch machinery in the runtime. For FixedTypeInfo, provide implementations that use the fixed spare bits derived for a type. Leave them unimplemented for dynamic types for now.

Swift SVN r7972