IRGen: Key dependent generic TypeInfos by bound generic type.
Ensure that the forward declaration and TypeInfo for generic struct and union types with non-fixed layout is keyed by a particular bound generic type instance rather than the unbound type. This ensures that the correct instance lands in the WitnessSizedTypeInfo so it can be used to look up the value witness table.

Swift SVN r8002