Add a new (temporary and hidden) commandline option: -enable-top-level-definite-init.
When enabled, this causes references to global variables in the 
top_level_code function to be emitted as:

  %0 = global_addr #x : $*Int64
  %1 = mark_uninitialized %0 : $*Int64

where all local uses of "x" use the %1 value.  This representation will
allow the definite initialization pass to know that "x" must be forced
to be initialized before the function returns or the value escapes (which
we don't have a representation for yet).



Swift SVN r8147