IRGen: Bind the destination BB argument for dynamic_lookup_br.
This is kind of gross because we want the BB argument to be lowered to an ObjCMethod abstraction rather than a concrete explosion of PHI nodes. For now, assume that there won't be other predecessors to a dynamic_method_br's destinations and clobber the BB argument's PHI node lowering, replacing it with an ObjCMethod. This is hacky but unblocks Doug's dynamic lookup work, and avoids having to create an ObjCMethod-to-explosion thunk in the (currently 100%, probably always 99.99%) case where the branches have single predecessors.

Swift SVN r8150