Introduce TypeBase::getTypeOfMember() to produce the type of a member access.
This routine provides an AST-level operation that takes a base type
and a member, then computes the type the member will have when
accessed through that base, substituting (e.g.) generic arguments for
the archetypes in the type of the member. It is a cleaned up, slightly
generalized, and slightly more correct formulation of
TypeChecker::substMemberTypeWithBase(), which is now just a shim that
invokes TypeBase::getTypeOfMember().

Addresses <rdar://problem/14778376>.


Swift SVN r8198