Type-check both sides of a PatternBindingDecl within the same constraint system.
Tweak the tuple conversion rule so that labeled elements can lose
their label (but not change it or get passed as a vararg) if the
label wasn't written immediately in a tuple literal.  This permits
  var (a,b) = fn_returning_labeled_tuple()
to successfully type-check, as well as the analogous assignment.

This implies an unfortunate number of test changes, including some
QoI regressions, which I tried to limit as much as I could.

Swift SVN r8438