Code completion: don't show [byref] on 'self'
Of course, in structs 'self' is a reference to the value, but this is so
obvious and natural, that calling out this [byref] in code completion results
every single time adds nothing but syntactic noise.

We still show [byref] in other cases, for example, when we complete references
to user-written function parameters.


Swift SVN r8464