Code completion: don't crash when completing in constructors/destructors placed in contexts where they are not allowed
We used to throw away the ConstrutorDecl/DestructorDecl, but code completion
captured the pointer as a DeclContext for the code completion point.  Then code
completion passed that AST node to type checker, which crashed.  Now we
properly mark the decl as invalid, pass it to the type checker, and have the
type checker set its type to ErrorType.


Swift SVN r8466