[weak] variables have Optional type.
This contains one significant difference from the design:
previously, you declared:
  var [weak] x : NSObject
and the type of 'x' was implicitly Optional.  This change
in the formal type proved problematic in a number of ways,
and Joe made the excellent suggestion to eliminate it:
so now you have to declare:
  var [weak] x : NSObject?
and [weak] is just a kind of modifier on the storage of the
optional value.  You could argue that this is somewhat
pedantic, but I think it restores a lot of consistency to
the model.

Swift SVN r8538