Bind extensions in all visible modules, rather than just the current TU.
This is still not fully correct: in theory a source module that isn't the
current TU could reference something in an extension that only it can see.
However, doing this properly means allowing decls with the same name to
coexist in a nominal, then choosing the right one based on which modules
are currently visible. This is more complexity than we want right now.

Swift SVN r8642