Record known protocols without calling the type checker
Right now the type checker is responsible for marking known protocols in the
imported AST for 'swift' module.  If we don’t run the type checker, known
protocols are not marked correctly.

This patch moves the code to mark known protocols from libSema to libAST and
calls it when we load the standard library module.

Motivation: Swift tools might want to deserialize a module and inspect the AST
(for example, print it).  There is no need to invoke the type checker in this
case, and we don’t even have a TU.


Swift SVN r8704