Sema: Resolve and type-check NominalTypePatterns in cases.
If we see 'T(...)' in a pattern context and T references a type by name, resolve it to a NominalTypePattern. Teach coerceToType how to look up properties in the type and coerce subpatterns to their types. For now, we only support a 1:1 mapping from the subject type to the declared type. An obvious extension would be to introduce an implicit IsaPattern if a cast can provide the pattern's type.

Swift SVN r8772