Resolve dependent member types to specific associated types.
Once we've determined all of the constraints for a generic function
signature, check the types in the signature again to resolve any
dependent member types to specific associated types. This gives us a
more sensible, type-checked dependent type for the signature.

We still throw away the resulting type, so this is still useless.



Swift SVN r8846