Disable the diagnostics level "stack to SSA" pass, which is almost completely subsumed by definite initialization.
The single case that definite initialization misses (as illustrated by the 
testcase) are opportunities to eliminate the load feeding the final release 
that is produced by the box-to-stack pass.

I'm fine ignoring this one minor, because this doesn't matter for diagnostic 
passes, and we'll eventually have a general load-elimination pass if this is 
important.



Swift SVN r8995